+++ gen/openapi/internal/rust/src/models/matchmaker_lobbies_find_request.rs
@@ -21,6 +21,8 @@ pub struct MatchmakerLobbiesFindRequest {
     pub prevent_auto_create_lobby: Option<bool>,
     #[serde(rename = "regions", skip_serializing_if = "Option::is_none")]
     pub regions: Option<Vec<String>>,
+    #[serde(rename = "tags", skip_serializing_if = "Option::is_none")]
+    pub tags: Option<::std::collections::HashMap<String, String>>,
     #[serde(rename = "verification_data", default, with = "::serde_with::rust::double_option", skip_serializing_if = "Option::is_none")]
     pub verification_data: Option<Option<serde_json::Value>>,
 }
@@ -32,6 +34,7 @@ impl MatchmakerLobbiesFindRequest {
             game_modes,
             prevent_auto_create_lobby: None,
             regions: None,
+            tags: None,
             verification_data: None,
         }
     }