+++ svc/pkg/mm/worker/src/workers/lobby_find/mod.rs
@@ -176,6 +176,7 @@ async fn worker(ctx: &OperationContext<mm::msg::lobby_find::Message>) -> GlobalR
 					.iter()
 					.filter_map(|player| player.client_info.clone())
 					.collect::<Vec<_>>(),
+				tags: &ctx.tags,
 
 				lobby_groups: &lobby_group_config.lobby_groups,
 				lobby_group_meta: &lobby_group_config.lobby_group_meta,
@@ -233,6 +234,7 @@ async fn worker(ctx: &OperationContext<mm::msg::lobby_find::Message>) -> GlobalR
 			query,
 			lobby_group_config: &lobby_group_config,
 			auto_create_lobby_id,
+			tags: &ctx.tags,
 		},
 	)
 	.await?