+++ svc/pkg/mm/worker/tests/lobby_create.rs
@@ -44,6 +44,7 @@ async fn setup(ctx: &TestCtx) -> Setup {
 				max_players_direct: 10,
 				max_players_party: 12,
 				listable: true,
+				taggable: false,
 
 				runtime: Some(backend::matchmaker::lobby_runtime::Docker {
 					build_id: build_res.build_id,
@@ -104,6 +105,7 @@ async fn setup(ctx: &TestCtx) -> Setup {
 				max_players_direct: 10,
 				max_players_party: 12,
 				listable: true,
+				taggable: false,
 
 				runtime: Some(backend::matchmaker::lobby_runtime::Docker {
 					build_id: build_res.build_id,